starting migration of subject screen

+++ app/src/main/java/biz/eventually/atpl/data/network/SubjectNetwork.kt
@@ -3,4 +3,4 @@ package biz.eventually.atpl.data.network
 /**
  * Created by thibault on 21/03/17.
  */
-data class SubjectNetwork(val id: Int, val name: String, val topics: List<TopicNetwork>?)
+data class SubjectNetwork(val id: Long, val name: String, val topics: List<TopicNetwork>?)
